修复：Ring相关

+++ src/Application.Core/Game/Players/IPlayer.cs
@@ -392,7 +392,7 @@ namespace Application.Core.Game.Players
         HashSet<NewYearCardRecord> getReceivedNewYearRecords();
         int getRelationshipId();
         int getRemainingSp();
-        Ring? getRingById(int id);
+        Ring? getRingById(long id);
         RockPaperScissor? getRPS();
         int getSavedLocation(string type);
         string? getSearch();
